refactor(submit) extract post submission action
