+++ src/Utils/Startup/ServiceCollectionExtensions.cs
@@ -90,6 +90,7 @@ using StockportGovUK.NetStandard.Gateways.StreetService;
 using StockportGovUK.NetStandard.Gateways.VerintService;
 using form_builder.Services.FormAvailabilityService;
 using form_builder.Providers.FileStorage;
+using form_builder.SubmissionActions;
 
 namespace form_builder.Utils.Startup
 {
@@ -266,6 +267,13 @@ namespace form_builder.Utils.Startup
         {
             services.AddSingleton<IStreetProvider, FakeStreetProvider>();
             services.AddSingleton<IStreetProvider, ServiceStreetProvider>();
+
+            return services;
+        }
+
+        public static IServiceCollection ConfigurePostSubmissionActions(this IServiceCollection services)
+        {
+            services.AddSingleton<IPostSubmissionAction, PostSubmissionAction>();
 
             return services;
         }
